fix(DeleteHosp,Listagem):arrumando o dialog e o data-set de deleção e listagem de hóspede

+++ scripts/DeleteHospConnection.js
@@ -1,33 +1,35 @@
 let deleteCpf = null;
 
-tableBody.addEventListener("click", async (e) => {
-    if(e.target.classList.contains("trashDel")){
-        deleteCpf = e.target.getAttribute("data-set")
-        const dialogDeleteHosp = document.querySelector(".deleteGuest")
-        dialogDeleteHosp.showModal();
+const tableBody = document.getElementById("guest-table-body");
+const dialogDeleteHosp = document.querySelector(".deleteGuest");
+const deleteCancelBtn = document.querySelector(".cancelDeleteHospede");
+const deleteConfirmBtn = document.querySelector(".confirmProssesForDeleteHopede");
 
-       const deleteCancelBtn = document.querySelector(".cancelDeleteHospede")
-   deleteCancelBtn.addEventListener("click", () =>{
+
+tableBody.addEventListener("click", (e) => {
+  if (e.target.classList.contains("trashDel")) {
+    deleteCpf = e.target.getAttribute("data-set");
+    dialogDeleteHosp.showModal();
+  }
+});
+
+
+deleteCancelBtn.addEventListener("click", () => {
+  dialogDeleteHosp.close();
+});
+
+
+deleteConfirmBtn.addEventListener("click", async () => {
+  if (!deleteCpf) return;
+
+  try {
+    await fetch(`${config.API_URL}/new-guest/delete-guest/${deleteCpf}`, {
+      method: "DELETE",
+      headers: { "Content-Type": "application/json" },
+    });
     dialogDeleteHosp.close();
-   })
-    }
-})
-
-
- const deleteConfirmBtn = document.querySelector(".confirmProssesForDeleteHopede")
-    deleteConfirmBtn.addEventListener("click", async () => {
-         try{
-            
-            await fetch(`${config.API_URL}/new-guest/delete-guest/${deleteCpf}`, {
-                method: "DELETE",
-                headers: { 'Content-Type': 'application/json'}
-            })
-        } catch(error){
-            console.error("Erro ao deletar hóspede:",  error)
-        }
-         
-        
-    })
-
-
-   
+    window.location.reload();
+  } catch (error) {
+    console.error("Erro ao deletar hóspede:", error);
+  }
+});
